backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.13.1
python-dotenv==1.0.0
email-validator==2.1.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pyjwt==2.8.0
imapclient==3.0.1
beautifulsoup4==4.12.2
python-multipart==0.0.6
fastapi-cors==0.0.6 